add user auth frontend and protection to routes at backend

+++ backend/controllers/todoControllers.js
@@ -3,7 +3,9 @@ const Todo = require('../models/todoModel')
 const mongoose = require('mongoose')
 
 const getTodos = async (req,res) => {
-    const todos = await Todo.find()
+    const { user_id } = req.body
+
+    const todos = await Todo.find({user_id})
 
     if(!todos) {
         res.status(200).json({error: "No todos"})
@@ -14,13 +16,14 @@ const getTodos = async (req,res) => {
 
 const getTodo = async (req,res) => {
     const id = req.params.id
+    const { user_id } = req.body
 
     if(!mongoose.Types.ObjectId.isValid(id)) {
         res.status(404).json({error: "No such todo exists"})
         return
     }
 
-    const todo = await Todo.find({_id: id})
+    const todo = await Todo.find({_id: id, user_id})
 
     if(!todo) {
         res.status(404).json({error: "No such todo exists"})
@@ -31,10 +34,23 @@ const getTodo = async (req,res) => {
 }
 
 const createTodo = async (req,res) => {
-    const {name, desc} = req.body
+
+    const {name, desc, user_id } = req.body
+    const emptyFields = []
+
+    if(!name) {
+        emptyFields.push('name')
+    }
+    if(!desc) {
+        emptyFields.push('desc')
+    }
+
+    if(emptyFields.length != 0) {
+       return res.status(400).json({error: 'Please fill all the fields', emptyFields})
+    }
 
     try {
-        const todo = await Todo.create({name, desc})
+        const todo = await Todo.create({name, desc, user_id})
         res.status(200).json(todo)
     }
     catch(err) {
@@ -63,13 +79,14 @@ const updateTodo = async (req,res) => {
 
 const deleteTodo = async (req,res) => {
     const id = req.params.id
+    const { user_id } = req.body
 
     if(!mongoose.Types.ObjectId.isValid(id)) {
         res.status(404).json({error: "No such todo exists"})
         return
     }
 
-    const todo = await Todo.findOneAndDelete({_id: id})
+    const todo = await Todo.findOneAndDelete({_id: id, user_id})
 
     if(!todo) {
         res.status(404).json({error: "No such todo exists"})
